arithmetic fuzzy
