+++ include/eve/module/arithmetic/fuzzy/impl/is_equal.hpp
@@ -7,19 +7,16 @@
 //==================================================================================================
 #pragma once
 
+#include <eve/module/core.hpp>
 #include <eve/concept/value.hpp>
 #include <eve/concept/compatible.hpp>
 #include <eve/module/arithmetic/constant/eps.hpp>
 #include <eve/detail/apply_over.hpp>
 #include <eve/detail/implementation.hpp>
 #include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/abs.hpp>
-#include <eve/function/bit_cast.hpp>
-#include <eve/function/convert.hpp>
-#include <eve/function/dist.hpp>
-#include <eve/function/is_less_equal.hpp>
-#include <eve/function/fuzzy.hpp>
-#include <eve/function/nb_values.hpp>
+#include <eve/module/arithmetic/regular/abs.hpp>
+#include <eve/module/arithmetic/regular/dist.hpp>
+#include <eve/module/arithmetic/regular/is_less_equal.hpp>
 #include <eve/traits/as_logical.hpp>
 
 namespace eve::detail
