+++ include/eve/module/arithmetic/fuzzy/impl/is_less_equal.hpp
@@ -8,12 +8,10 @@
 #pragma once
 
 #include <eve/detail/implementation.hpp>
-#include <eve/function/fuzzy.hpp>
-#include <eve/function/saturated.hpp>
-#include <eve/function/abs.hpp>
-#include <eve/function/is_less_equal.hpp>
-#include <eve/function/saturated/next.hpp>
-#include <eve/function/saturated/add.hpp>
+#include <eve/module/arithmetic/regular/abs.hpp>
+#include <eve/module/arithmetic/regular/is_less_equal.hpp>
+#include <eve/module/arithmetic/saturated/next.hpp>
+#include <eve/module/arithmetic/saturated/add.hpp>
 #include <eve/concept/value.hpp>
 #include <eve/concept/compatible.hpp>
 #include <eve/module/arithmetic/constant/eps.hpp>
